restructured functions for building agent
Split loadAgentParams.m into two functions; one defines the parameter settings of the agent, and the second uses these parameter settings to construct the computational modules of the agent. This decomposition is easier to parse and modify, and it matches the organization of functions for building the environment and trial structures.