modified bounding of trajectory by arena
previously, the trajectory was bounded during execution (in executeTrajectory.m); as a result, the planned and executed likelihoods always differed based on both obstacle and boundary encounters. Now, the trajectory is bounded during the planning/optimization phase, based on the idea that the agent knows the overall size of the arena. This modification means that the planned and executed likelihoods differ only base don unexpected encounters with obstacles, and not "expected" encounters with arena boundaries