redefined reference for trajectory generator
Modified the reference point for defining the heading offset, such that positive and negative deviations are defined relative to the vector separating two anchors. This allowed for an easy fix of the boundary conditions at +/- pi. This also required modifying the method for diverting trajectories around obstacles to ensure that planned diverted paths don't intersect the obstacle