added option to cache belief
the agent can now store its current belief in a memory cache, and sample from the memory cache to guide the sampling of new anchor points. To enable this, the agent maintains and updates a hierarchical belief about the current context, and about the target location within a context.

This update includes:
(1) options for allowing/disallowing caching, setting the max size of the cache, and specifying the method used to sample from the cache.
(2) a new function 'sampleContext' that uses the context prior to sample a belief map from the memory cache.
(3) a modification of 'updateBelief' to update both context prior and location prior.
(4) a modification of some storage variables.